% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Effective sample sizes from a cMAIC fit}
\usage{
effective_sample_size(object, ...)
}
\arguments{
\item{object}{A \code{cpaic_maic} object.}

\item{...}{Unused.}
}
\value{
A named numeric vector of effective sample sizes per IPD study.
}
\description{
Effective sample sizes from a cMAIC fit
}
