% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_diagnostics.R
\name{waic.cpaic_mlnmr}
\alias{waic.cpaic_mlnmr}
\title{Widely applicable information criterion}
\usage{
\method{waic}{cpaic_mlnmr}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=cmlnmr]{cmlnmr()}} fit.}

\item{...}{Passed to \code{\link[loo:waic.matrix]{loo::waic.matrix()}}.}
}
\value{
A \code{waic} object from the \code{loo} package.
}
\description{
Widely applicable information criterion
}
