% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_diagnostics.R
\name{loo.cpaic_mlnmr}
\alias{loo.cpaic_mlnmr}
\title{Pareto-smoothed importance sampling leave-one-out cross-validation}
\usage{
\method{loo}{cpaic_mlnmr}(x, ...)
}
\arguments{
\item{x}{A \code{\link[=cmlnmr]{cmlnmr()}} fit.}

\item{...}{Passed to \code{\link[loo:loo.matrix]{loo::loo.matrix()}}.}
}
\value{
A \code{psis_loo} object from the \code{loo} package.
}
\description{
Pareto-smoothed importance sampling leave-one-out cross-validation
}
